Created HomeFeedService Refactored HomeViewController to use HomeFeedService
Created ActivityService
Refactored ActivityViewController and ActivityViewModel to use ActivityService

Completely changed the way new post is made. There's now a cool animation to show a notification which displays progress of making a post and seamlessly transforms into a post itself once it's posted.

Created PostWithCommentsViewController to display a post with its comments when user selects activity .postLiked or .postCommented notification.
PostWithCommentsService and ViewModel were created to accompany it.

Lots of other small changes.